Remove old philosophy page
