+++ web/src/pages/demo/index.tsx
@@ -5,7 +5,7 @@ import Page from "../../components/Page"
 import Section, { SectionTitle } from "../../components/Section"
 import Navigation from "../../components/Navigation"
 import FAQs, { FAQ } from "../../components/FAQs"
-import PhilosophyV2 from "../../components/PhilosophyV2"
+import PhilosophyV2 from "../../components/Philosophy"
 import ContactForm from "../../components/ContactForm"
 import Panel from "../../components/Panel"
 import Footer from "../../components/Footer"
