+++ web/src/pages/edinburgh/index.tsx
@@ -8,7 +8,7 @@ import Navigation from "../../components/Navigation"
 import Cover, { CTADown } from "../../components/Cover"
 import IntroStats from "../../components/IntroStats"
 import FAQs, { FAQ } from "../../components/FAQs"
-import PhilosophyV2 from "../../components/PhilosophyV2"
+import PhilosophyV2 from "../../components/Philosophy"
 import ContactForm from "../../components/ContactForm"
 import SocialIcon from "../../components/SocialIcon"
 import Panel from "../../components/Panel"
