+++ web/src/pages/index.tsx
@@ -6,7 +6,7 @@ import Section, { SectionTitle } from "../components/Section"
 import { TopNavigation } from "../components/Navigation"
 import Cover, { CTADown } from "../components/Cover"
 import IntroStats from "../components/IntroStats"
-import PhilosophyV2 from "../components/PhilosophyV2"
+import PhilosophyV2 from "../components/Philosophy"
 import ContactForm from "../components/ContactForm"
 import Footer from "../components/Footer"
 import SocialIcon from "../components/SocialIcon"
